refactor(connect): single-source the delivery.http resolver in afps-runtime
`resolveHttpDelivery` + the alias-aware `readCredentialField` + `ALIAS_MAP` +
`AUTH_TYPE_HTTP_DEFAULTS` existed twice: once in `@appstrate/connect`
(platform) and a hand-maintained "portable mirror" in
`@appstrate/afps-runtime` (the standalone CLI). The two had already drifted —
connect keyed defaults in camelCase with a 12-entry alias map, afps-runtime in
snake_case with a 3-entry one.

afps-runtime is the dependency-free bottom layer and `@appstrate/connect`
already depends on it (it re-exports `matchesAuthorizedUriSpec` from there),
so the canonical copy now lives in `packages/afps-runtime/src/resolvers/
http-delivery.ts` and connect re-exports it. Standardised on camelCase
defaults + the 12-entry alias map (the superset — `readCredentialField`
resolves either casing, so this strictly widens alias coverage for the CLI
without changing platform behaviour).

- new `http-delivery.ts`: `resolveHttpDelivery`, `readCredentialField`,
  `ALIAS_MAP`, `HttpDeliveryConfig`, `HttpDeliveryPlan`; exported from the
  resolvers barrel.
- afps-runtime `integration-api-call.ts`: drop the mirror
  (`resolveLocalHttpDelivery` + dup defaults/alias-map/reader), call the
  canonical `resolveHttpDelivery`.
- connect `integration-credentials.ts`: drop its definitions, re-export the
  canonical ones so every `@appstrate/connect` consumer keeps its import path.

Gate green. 1159 pass across afps-runtime + connect + sidecar + the DB-backed
credential-proxy injection suite (all exercise the resolver).
